requirements.txt
pyarrow
python-dotenv
snac
peft
soundfile